+++ build-scripts/build.sh
@@ -50,6 +50,8 @@ fi
 npm run build
 cp -r themes output/themes
 cp -r debug output/debug
+mkdir output/lib
+cp -r lib/basisu output/lib/basisu
 
 #set +e
 #npm run start:pack & npm run cy:local