+++ dev/scripts/gmi.js
@@ -34,6 +34,11 @@ var GMI = function (options, embedVars, gameDir) {
 			return gameDir;
 		},
 	});
+	Object.defineProperty(GMI.prototype, "info", {
+		get: function () {
+			return { gameDir };
+		},
+	});
 	Object.defineProperty(GMI.prototype, "environment", {
 		get: function () {
 			return env;
@@ -236,13 +241,15 @@ var GMI = function (options, embedVars, gameDir) {
 			var achievementsCloseButton = document.createElement("input");
 			achievementsCloseButton.type = "button";
 			achievementsCloseButton.value = "Click here to close.";
+			var achievementsData = achievementsCloseButton.dataset
+			achievementsData.testid= "close-achievements";
 
 			achievementsDiv.appendChild(achievementsCloseButton);
-
 			achievementsCloseButton.addEventListener("click", function () {
 				document.body.removeChild(achievementsDiv);
 			});
 			document.body.appendChild(achievementsDiv);
+			achievementsCloseButton.focus();
 			return true;
 		}
 		return false;