+++ src/core/accessibility/accessible-dom-element.js
@@ -3,7 +3,7 @@
  * @author BBC Children's D+E
  * @license Apache-2.0
  */
-import crel from "../../../lib/crel.es.js";
+import crel from "/node_modules/crel/crel.es.js";
 import fp from "../../../lib/lodash/fp/fp.js";
 
 const keyUp = options => event => {