+++ src/core/browser.js
@@ -21,13 +21,9 @@ export const getBrowser = () => {
 
 	const isSilk = name === "Amazon Silk" || isKindleWebView;
 
-	//Ipad 2 faster in Canvas. Force bool as Bowser returns undefined if browser doesn't match
-	const forceCanvas = isSilk || Boolean(browserInfo.satisfies({ safari: "<10" }));
-
 	return {
 		name,
 		version,
-		forceCanvas,
 		isSilk,
 	};
 };