+++ src/core/layout/dom-text.js
@@ -24,20 +24,40 @@ const defaultStyle = {
 	position: "absolute",
 };
 
-const getTextNodes = text => text.split("\n").map(line => document.createTextNode(line));
+const getTextNodes = text => text.split("\n").map(line => document.createTextNode(line.trim()));
 const addBreaks = (el, i, arr) => [el].concat(i !== arr.length - 1 ? [document.createElement("br")] : []);
+
+let uid = 0;
 
 class DomText {
 	constructor(text, newConfig) {
 		const config = { ...defaultConfig, ...newConfig };
 		this.el = document.createElement("div");
+		this.el.id = `domtext-${uid++}`;
+		this.el.dataset.text = text;
 		this.setStyle({ ...defaultStyle, ...config.style });
-
 		this._textNodes = [];
 		this.setText(text);
 
 		this.setAlignment(config.align);
 		this.setPosition(config.position.x, config.position.y);
+	}
+
+	addOuterStroke(size, color) {
+		this.inlineStyleSheet = document.createElement("style");
+		const css = `
+		  #${this.el.id}::after {
+		  content: attr(data-text);
+		  position: absolute;
+		  left: 0;
+		  top:0;
+		  -webkit-text-stroke: ${size * 2}px ${color};
+		  width: ${this.el.style.width};
+		  z-index:-1;
+		  font-size: 1em;
+		}`;
+		this.inlineStyleSheet.appendChild(document.createTextNode(css));
+		gelDom.current().appendChild(this.inlineStyleSheet);
 	}
 
 	setText(newText) {
@@ -63,6 +83,7 @@ class DomText {
 
 	destroy() {
 		this.el.remove();
+		this.inlineStyleSheet?.remove();
 	}
 }