+++ src/core/loader/phaser-defaults/get-phaser-defaults.js
@@ -26,7 +26,7 @@ export const getPhaserDefaults = config => {
 	return {
 		...getBaseDefaults(),
 		...{
-			type: browser.forceCanvas ? Phaser.CANVAS : Phaser.AUTO,
+			type: Phaser.AUTO,
 			transparent: browser.isSilk, // Fixes silk browser flickering
 			parent: getContainerDiv(),
 			scene,