+++ test/mock/mock-scene.js
@@ -3,6 +3,15 @@
  * @author BBC Children's D+E
  * @license Apache-2.0
  */
+const context = {
+	measureText: () => 100,
+	save: () => {},
+	scale: () => {},
+	translate: () => {},
+	fillText: () => {},
+	restore: () => {},
+};
+
 export const mockBaseScene = () => ({
 	sys: {
 		anims: { once: () => {} },
@@ -13,6 +22,7 @@ export const mockBaseScene = () => ({
 			get: () => ({ get: () => ({ cutWidth: 0, cutHeight: 0 }) }),
 			addCanvas: () => ({
 				get: () => ({ source: {}, resolution: {}, setSize: () => {} }),
+				context,
 			}),
 		},
 	},